requirements.txt
pyyaml
argparse
pathlib
numpy
scipy
plotly
torch
pandas
pyarrow
matplotlib
scikit-learn
torch-geometric
wandb